requirements.txt.txt
torch
torchvision
Pillow

"Firstchange location according to the device, Then run mobilenetv2.py, and then run mobilenetv2run.py"